// src/app/models/grading/all-test-type-details-viewmodel.ts
/**
 * BaseWebAPI
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


/**
 * AllTestTypeDetailsViewModel
 */
export interface AllTestTypeDetailsViewModel { 
    sortBy: string;
    sortOrder?: number;
    names?: Array<string>;
    codes?: Array<string>;
    questionCounts?: Array<number>;
    marksCounts?: Array<number>;
    pageNumber: number;
    pageSize: number;
}