// src/app/models/grading/test-type-view.ts
/**
 * BaseWebAPI
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


/**
 * UpdateTestTypeViewModel
 */
export interface TestTypeView { 
    /**
     * Gets or Sets of Id
     */
    id: number;
    /**
     * Gets or Sets of Name
     */
    name: string;
    /**
     * Gets or Sets of Code
     */
    code?: string;
    /**
     * Gets or Sets of QuestionCount
     */
    questionCount?: number;
    /**
     * Gets or Sets of MarksCount
     */
    marksCount?: number;
}